+++ OneWoW_AltTracker/OneWoW_AltTracker.lua
@@ -78,7 +78,7 @@ local function OnInitialize()
     end)
 
     local _ver = OneWoW:GetAddonVersion(ADDON_NAME)
-    OneWoW:RegisterLoadComponent("AltTracker", _ver, "/1wat")
+    OneWoW:RegisterLoadComponent("AltTracker", _ver, "/1wat", ADDON_NAME)
 end
 
 function OneWoW_AltTracker:ApplyTheme()